update unclutter
Don't clear the pointer focus when hiding the cursor. This was annoying
because if you e.g. had the cursor over a spoiler tag it becomes black
again after the timeout, or if the cursor was opening a popup by
hovering over a button the popup disappears after the timeout.

Not clearing the pointer focus requires adding a check when changing the
cursor image to not make the cursor reappear when it does. Skipping the
image change results in having the old cursor image if it changed while
the cursor was hidden, until it is changed again. Don't keep track of
the last cursor image for simplicity.