update unclutter
Remember the last cursor shape when restoring the cursor.

Remove the condition that returns early when a pointer constraint is
active because it is not necessary now that pointer focus is no longer
cleared.

I didn't call wlr_seat_pointer_notify_enter() while the cursor is hidden
to avoid restoring it, but this is no longer necessary now that
setcursor() and setcursorshape() don't change the cursor while it is
hidden: those calls were what restored the cursor.

By calling wlr_seat_pointer_notify_enter() even when the cursor is
hidden, we no longer have to manually restore pointer focus when the
cursor is restored.